VoteContainerとVoteをGeneratedImageContainerに移動
